requirements.txt
Django==4.2.7
djangorestframework==3.14.0
pymongo==4.6.1
